+++ core/tests/conftest.py
@@ -39,12 +39,6 @@ def _clear_test_artifacts(test_name: str) -> None:
         for key in list(cache.iterkeys()):
             del cache[key]
 
-    with local_persistent_storage("media_throughput") as c:
-        c.pop(test_name)
-
-    with local_persistent_storage("alt_media_throughput") as c:
-        c.pop(test_name)
-
     with local_persistent_storage("alt_media_fraction") as c:
         c.pop(test_name)